Drastically upgrade WebView.vue
The iframe in WebView now serves html scrapped from extensions with an
injected script that allows iframe to communicate its current url to its
parent. This allows Manga Desktop user to browse a source's website and
bookmark a series directly from said website.

The main purpose of this is to reduce the need of scrapping and creating
a complicated series filter when browse a source. This also allows the
removal of ads and clickbait scripts when html is being scrapped.